Add files via upload
Utilized Data Science and Machine Learning techniques to predict milk quality, ensuring the safety and excellence of dairy products. Developed a predictive model based on seven key features to maintain high-quality standards in the dairy industry. With a Kaggle-based dataset comprising 1074 rows and 8 columns, the model achieved an accuracy rate of approximately 91%.